Hotfix: correctly compute mUpdate notifications against the updated document (#1307)
# Description

When doing a `document:mUpdate`, real-time notifications are erroneously computed on the update query content, instead of the full updated document.

This is because for mUpdate we use Elasticsearch's `source` option, asking ES to return the updated document source, saving us from performing a `get` query for each updated document. But the returned source doesn't replace the `_source` property, which still contains only the update content. Instead, the full document source is put in `get._source`.

This PR correctly replaces _source with get._source before returning the result to the document controller.

# How to test

* Subscribe with the following filters: `{exists: 'foo'}`
* Create a document with the following body: `{foo: true, bar: 'baz'}`
* Use mUpdate to update it, **without** changing `foo`. Body example: `documents: [{_id: <documentId>, body: {bar: 'qux'}}]`

Without this fix, an `out` notification is sent, instead of an `in` one, because the notifier module thinks that the `foo` field has been removed and does not exist anymore.