Upgrade the WebSocket libraries (#1301)
# Context

First, some context that will help understand some of the changes made in that PR: the following explains why I went back on using `ws` instead of porting Kuzzle to  [µWebSockets.js](https://github.com/uNetworking/uWebSockets.js).

Kuzzle currently depends on the obsolete `uws` module. 
Initially I worked on replacing this module with its newer version (µWebSockets.js): but I discovered, once well into the massive needed refactor (and reading a lot of code because of the poor documentation), that despite its claim, this module does not fully implements the RFC. Most notably, [it doesn't allow a server to send PING requests](https://github.com/uNetworking/uWebSockets/issues/819) and, as always with this library's author, you have the choice of silently coping with the library flaws, or being insulted if you try to ask about the feature or -god forbids- if you are crazy enough to submit a PR.

So, I decided that continuing to use that author's libraries was a loss of time and even dangerous: I discovered about the unimplemented PING/PONG, but there might be other unimplemented RFC features that we might discover later, and we don't need any more bad surprises.
Sure, that library's performances are insane, but we already spent a lot of time in maintenance because of it.

That said, there were a couple of good ideas I picked up while refactoring the WebSocket layer that I decided to keep. So, you might see what seems "unrelated" changes that I _had_ to do with µWebSockets, and since the code was there, I kept it.

# Description

Because `uws` is obsolete and doesn't work on newer Node.js versions (it seems to work with Node.js 10 now, but not with Node.js 12), it was needed to use another library and, for the reasons explained above, it was decided to go back to the reliable and fully RFC compliant [ws](https://www.npmjs.com/package/ws).

Changes:
* Kuzzle now keeps track of activity for each opened sockets, to be able to detect stale or idle sockets when PING/PONG is disabled, which is a sensible choice for high amounts of clients connections (e.g. to terminate idle sockets even if the client is still alive). This can be configured using the new `idleTimeout` parameter in the kuzzlerc file
* WebSocket's `broadcast` method, used to send real-time notifications, has been massively optimized: since broadcasting notifications means sending very similar messages to a potentially very large number of sockets, that method now precomputes a RFC-6455 frame, and access directly on the low-level socket layer to send that frame, instead of asking to ws to compute a new frame for each target client socket. Benchmark results show 60k+ notifications/second sent to 10k registered sockets, on a mere laptop
* The WebSocket entrypoint now uses Set and Map instead of raw objects to keep track of connections and real-time subscriptions: this change has been made because Set/Map are much more suited (performance-wise) to hold data with constantly happening additions/removals 
* Checks made during startup to verify user configuration are now made using `assert`, as this makes much more sense since they are supposed to make Kuzzle stop

# Other changes

* the HTTP entrypoint has been refactored: *no changes were made to the logic*. Initially, a refactor was required because µWebSockets opens its own TCP socket, and manages its own HTTP routing, so I had to adapt that entrypoint accordingly and, along the way, I cleaned it up, fixed the outdated/erroneous JSDoc, and made other changes to make it much more readable and maintanable. I reverted the modifications needed by µWebSockets obviously, but I decided to keep the refactor: I just didn't want to throw my work out the window.