[fix] provide a copy of Kuzzle's configuration to plugins to prevent modifications (#1368)
Description
The configuration object provided to plugins' init functions is the one maintained by Kuzzle. Plugins modifying this object will also modify Kuzzle's object, altering the output of the server:getConfig API route result.

This PR makes sure that only a copy of the configuration file is provided to plugins.