Always set the content length for HTTP responses (#1375)
Before this PR we never set the Content-Length header for HTTP responses. This wasn't a problem, until one of our users made their plugin set one, and when a requesting client asked for a compressed response: the content length set by our user plugin couldn't be the same than the compressed payload, and this resulted in errors with, for instance, the SPDY protocol.

This PR makes Kuzzle always set the Content-Length header, overwritting it if it's already set.