Fix cache roles and profiles (#1433)
This PR fixes the way that we store a role and a profile after writing on it.
Before that, what was stored was the content of the request, and so was different from what was stored in ES.
While using getRole and searchRoles : result was different. searchRoles has the expected behavior since it don't load it from the cache.