Adapt Elasticsearch service for ES 7 (#1397)
## How to review

This PR is mainly about the rewrite of `elasticsearch.js` service, so you should review this file and the associated unit tests.  

### Why it's a mess ?

In a subsequent PR, I plan to refactor the usage of `InternalEngine` and the `Elasticsearch` service.  
We will have 2 instances of the `Elasticsearch` service:
 - `kuzzle.internalStorage` for internal indexes
 - `kuzzle.publicStorage` for public indexes
  
A new class named `IndexEngine` will replace the `InternalEngine`. This class will take `kuzzle.internalStorage` and restrict the methods to a specific index (eg: kuzzle internal index, plugin index, repositories, etc.)

So, there is a lot of code that didn't work for now and unit tests that are deactivated because I didn't want to make this PR bigger than it already is.

Also, I need to update the different controller.

## Elasticsearch service

The service use the new `@elastic/elasticsearch` client. The main difference is that now the response from Elasticsearch is inside the root `body` property of the object returned by the client.

Each method now takes several arguments instead of a single `request`.

Since the garbage collector will be removed in `2.x`, the code is much more simpler.

IndexCache checks to know if index/collection exists are removed because they will be done in the controllers.

 - `deleteByQuery`: **breaking change**  does not return the deleted documents ids
 - `truncateCollection`: delete the index, then recreate it with the same mappings
 - `getMapping`: **breaking change** return directly the mappings object
 - `refreshIndex`: **breaking change** renamed to `refreshCollection`
 - `setAutoRefresh`: **breaking change** method removed
 - `getAutoRefresh`: **breaking change** method removed
 - `bulk`: **breaking change** bulk is possible only on one collection
 - `m*`: **breaking change** return an object with `result` and `errors`

**breaking change** `truncateCollection`, `createCollection`, `updateMapping` does not return `acknowledged: true` anymore

## Other

  - collectionEmulator.js is now integrated in elasticsearch.js
  - Remove internalEngine