Refactor services and storage engines (#1436)
## Breaking changes

 - the configuration `internalEngine` is renamed `internalIndex`.

## Services

Our only services are `Elasticsearch` and `Redis`.  

Each service is duplicated for internal and public use:
```
kuzzle.services = {
  // Elasticsearch for internal indexes ('%')
  internalStorage,
  // Redis for internal usage (indexCache, scrollId, ...)
  internalCache,
  // Elasticsearch for public indexes ('&')
  publicStorage,
  // Redis for public usage (ms controller)
  publicCache
} 
```

The dynamic loading of services is removed and also the configuration saving in database.  
All services extends the `Service` base class which include an `init` method with a configurable timeout.  
They are all initialized within the `Services` (notice the `s` ;) ) class.

## Storage Engine

A storage engine is a class who allows to write data to the database.  

There are the `internalStorage` and `publicStorage` storage engine who are instance of the `Elasticsearch` service scoped to `public` or `internal` indexes.  
`publicStorage` is used by the controllers and `internalStorage` is indirectly used by Kuzzle itself for its internal index and for the plugins through the `accessors.storage` or `constructor.Repository`.

The `IndexEngine` class is a wrapper around a storage engine (`publicStorage` or `internalStorage`) scoped to an index.  

This class is used to write in the internal index or in plugin indexes.

Exemple usage:
```
// Instantiate a restricted storage for internal index 'my-index'
this.internalIndex = new IndexEngine('my-index', this.services.internalStorage);
// Create a document in my-index/my-collection
this.internalIndex.create('my-collection', { body: 'foobar' }, { id: 'document-id' });
```

The `kuzzle.internalEngine` core component has been replaced by `kuzzle.internalIndex` which is an instance of `IndexEngine` scoped to the `kuzzle` index.  

The `IndexEngine` class accepts an additional bootstrap class to initialize the corresponding index:

```
// Instantiate bootstraper class with the IndexEngine
const bootstrap = new InternalIndexBootstrap(kuzzle, kuzzle.internalIndex);
// Attach bootstraper class
kuzzle.internalIndex.bootstrap = bootstrap;
// Init method will bootstrap the index
kuzzle.internalIndex.init()
```