Access to the plugin repository using the lowercased manifest name (#1459)
# Description

Since #1430 it is now possible to set plugin names using uppercases, and the plugin key is still lowercased to better detect duplicates. 
There were a couple places in plugin strategies where the plugins repository was still accessed using the raw manifest name, instead of lowercasing it beforehand. 

This leads to TypeError exceptions + crash when that happens.