Plugins and protocols manifest files are now required (#1448)
In Kuzzle v1, it was possible to create plugins or protocols without a manifest.json file.
In that case, Kuzzle logged a warning explaining that it was assuming that the plugin/protocol was compatible with Kuzzle v1, and then fallbacked to the package.json file.
The warning also stated that this fallback will cease to exist in v2.

So. Kuzzle v2 will be released, and this obsoletes this fallback.
That PR removes it, making Kuzzle crash with a meaningful error if a plugin or a protocol does not feature a manifest.json file.