Check for SDK compatibility against current Kuzzle version (#1495)
## What does this PR do ?

Now Kuzzle will check the `volatile.sdkName` header to ensure that the current SDK is compatible against the current Kuzzle version.  

For Kuzzle v1 SDKs, the `sdkName` property is not present. So when the property is missing, we assume that the SDK is compatible.

For Kuzzle v2 SDKs, a compatibility matrix has been added. It works with simplified semver ranges.

Example  compat matrix:
```
{
  "js": "<7" // this version of Kuzzle is compatible with the JS SDKs prior to the 7.x.x version
}
```

Example `sdkName` volatile property: `js@7.4.2`