Fix cluster promise leak (#11)
# Description

Prerequisites to allow the cluster plugin to not leak promise, by making a few hook-only events compatible with pipes.

# Other changes

Solved a crash occuring during startup, when bootstrapping for the internal index or for the plugin dedicated indexes.
The crash happens when the lock cannot be created because it has already been created by another node, but then fetching the lock document to get its release time throws an uncaught 404 error because the lock has since been released. The exception is now caught and the bootstrapper will retry to acquire a lock a bit later when that happens.