Remove superfluous shard check when checking connection to the ES stack (#1528)
# Description

Start scripts used by docker images create an index named `%___tmp` in Elasticsearch to force shards to initialize. And then the script deletes that index.

This introduces a race condition in a cluster environment when multiple Kuzzle nodes start at roughly the same time: during its starting sequence, Kuzzle lists the existing indexes in ES and tries to apply its default mapping, but sometimes the fake index disappears during that moment, aborting the node's start sequence.

After discussing the issue with the team, it seems that this index creation was an old workaround made for dev environment: it is unnecessary for prod environments, and probably for dev ones too.

# Why merging on master?

This PR only updates docker scripts, and it's beneficial to re-generate our current images with that fix.