Requests rate limit per user (#1546)
Description

Introduce configurable requests rate limit:

    - requests rate limit are per node, per second and per user. This means that a rate limit imposed to the anonymous profile will impact all anonymous requests, since those are all attached to the same user id (this behavior has been documented)
    - the rate limit is a new profile property, meaning that it can easily target only a selection of users if needed
    - unless specified otherwise, the rate limit of a profile is set to 0 (= no limit)
    - when restricting rights (e.g. using the createFirstAdmin API route), rate limits are imposed as follows: admin: no limit, default: 10, anonymous: 200
    - the auth:logout API route is never restricted when requested by authenticated users (since it has no sense that anonymous users execute that route, it stays limited for them)
    - the auth:login API route is NOT impacted by the rate limit
    - requests exceeding the rate limit are rejected with the new TooManyRequestsError error class introduced in kuzzle-common-objects@4.1.0 (error status: 429).

The auth:login API route is a special case: we do not want it to be unavailable to legitimate users if non-authenticated ones flood Kuzzle in the hope of preventing users to log in. So it is excluded from the rate limit system.
Instead, a new property has been added in the kuzzlerc configuration file: loginsPerSecond. This controls how many logins per second can be performed by any single network socket (not user accounts).
By default, it is set to 1, as I don't see how allowing more than 1 login per second and per network connection is beneficial.
This limit cannot be set to 0, there is no "no limit" setting for that one.
Other changes

    The funnel now rejects requests without a controller or without an action immediately, instead of waiting for a slot to open and checking rights, as such requests could never be executed

Boyscout

    asyncify methods