Fix settings loading fail on opened indexes (#1586)
## What does this PR do ?

When you create/update a collection, if the action fail then the index settings are rollback.  
But some of the original settings can not be set when the index is open. 

So when a collection create/update fail for some reason (mapping change or whatever), Kuzzle throw this error: `Can't update non dynamic settings [[index.number_of_shards]] for open indices [[%kuzzle.profiles/hqwAn8ztQeuWwan5K3Y0Fw]]`

 - this is useless because we always rollback, even when settings didn't changed 
 - some settings need the index to be closed, so it's safer to always close it before applying settings

Not every index setting change need it to be closed but with this fix we ensure this very annoying bug won't happen again.