Fix API key expiration on realtime subscription (#1619)
API key can never expire, so the expiresAt property of the token is -1.

When a realtime subscription is open, Kuzzle checks regularly if the token associated to the subscription is still valid. This check compare the current date with the expiresAt property so a negative value is considered as an expired token.