Fix: invalid context on plugin pipes declared with a function name (#1631)
# Description

This is an emergency patch, fixing a behavior broken by #1613 
Plugins can attach either functions or function names to pipe events. 

When using function names, plugin developpers have no way of controlling the context used to invoke their functions, and Kuzzle expects the function to exist within the plugin main object (this is verified with assertions when plugins are loaded by Kuzzle).

With Kuzzle 2.2.0, plugin pipes defined using function names are invoked with an undefined context, breaking existing implementations.

This PR fixes that by binding the plugin pipe function to its object of origin, when loading the plugin. 

This is only applied to plugin pipes defined with function names: when using functions, plugin developpers are responsible of the function's context, as Kuzzle has no way of knowing in what object the function belongs.

# How to test

Load a plugin with a pipe function declared using a function name, and use `this` in the function.

Example:

```js
'use strict';

class Foo {
  constructor() {
    this.foo = 'bar';
  }

  async init (config, context) {
    this.config = config;
    this.context = context;
    this.pipes = {
      'server:beforeNow': '_lol'
    };
  }

  async _lol (rq) {
    console.log('============= LOL: ', this.foo);
    return rq;
  }
}

module.exports = Foo;
```

Without this fix, this fails with a `PluginImplementationError` stating that the `foo` property cannot be read from `undefined`.
With the fix, this pipe runs as expected.

# Other changes

* Simplify unit tests a bit