Prevent unhandled rejection in plugin hooks (#1621)
## What does this PR do ?

If a handler attached to a hook rejects, then we end up with an unhandled rejection.

This PR ensure that all hook errors are properly handled. Also, hook errors now triggers the new `hook:onError` event.

If a hook on `hook:onError` fail, it will emit the `plugin:hook:loop-error` who is listened by the plugin manager to log the error.

### How should this be manually tested?

Listen to a plugin event with an async hook and throw an error. 

### Other changes
 
 - add functional test on hook