Allows to use realtime subscriptions in plugins (#1645)
## What does this PR do

This PR allows to use realtime subscriptions in plugin with the embedded SDK.

Subscription should be done in the `init` method or in a hook to the `core:start:before` event.  

:book: https://deploy-preview-1645--doc-kuzzle.netlify.app/plugins/plugin-context/accessors/sdk/

```js
class MyPlugin {
  async init (config, context) {
    this.sdk = context.accessors.sdk;


    await this.sdk.realtime.subscribe('index', 'collection', {}, notif => {
      console.log(notif.result);
    });
  }
}
```

### Behavior

When you subscribe to notification in the `init` method or in the `core:kuzzleStart` event, then every node make a subscription.

#### with `replicate: false`
   
When a request triggering a notification is received by one of the node, only his own subscription is triggered.  
It allows to send an email from a real-time subscription (you don't want each node to send an email obviously)

#### with `replicate: true`

When a request triggering a notification is received, then every node trigger his own subscription.  
It allows to maintain RAM cache for example. (eg: indexCache, repositories) 

### How it works

A new `InternalProtocol` has been created. This protocol only have one connection with the router.  

The `FunnelProtocol` used by the plugin SDK send every request with the `InternalProtocol` connection ID so when the notifier want to transmit realtime notification for a subscription made with the internal SDK it's sended to the `InternalProtocol`.

The `InternalProtocol` then send the notification through Kuzzle internal event system and the `FunnelProtocol` is listening to it and then it dispatch it to the SDK object with the `roomId` using the same mechanism as for every SDK protocol.

To avoid replicating the notification on every cluster node, a `cluster` option has been introduced in the channels description (and as an argument of the `realtime:subscribe` method). This argument is `true` by default.

The `HotelClerc` will not emit cluster subscription synchronization related events for channels that have the `cluster` option set to `false`.   
So a subscription made in a plugin with `cluster: false` will not be propagated to other node (each node will only be aware of his own local subscription).

The `Notifier` methods `_notifyDocument` and `_notifyUser` can be called either from the node that is actually triggering the notification (from `notifyDocument` and `notifyUser`) or from the cluster code when the `core:notify:document` or `core:notify:user` are emitted.  
When the methods are called from the primary node then an additional parameter is passed to `_notify*` methods and then it will also trigger channels that have the `cluster: false` option.  
When they are called from the cluster, this additional parameter is not present and so channels with the `cluster: false` option are not triggered.

### Known limitations

If a subscription is made inside a controller method without the `replicate: true` option, then the subscription will only be triggered when this node receive a request meant to trigger a realtime notification.  

In fact it could be considered harmful to subscribe to notifications at runtime because the notification will only be present on one node so if this node is shutdown or unavailable then the subscription will never be triggered.

IMHO runtime subscription should be avoided until we don't have a subscription propagation mechanism for app/plugins

### How should this be manually tested?

First, you will have to declare a plugin who subscribe in the `init` method.

Then you can run a cluster with the following commands in three different terminals:
```bash
# term 1
$ WITHOUT_KUZZLE docker-compose up

# term 2
$ kuzzle_plugins__cluster__privileged=true kuzzle_server__port=7512 NODE_ENV=development DEBUG=kuzzle:*,-kuzzle:events*,-kuzzle:entry-point:protocols:websocket,-kuzzle:funnel,-kuzzle:entry-point:protocols:http,-kuzzle:http:router,-kuzzle:plugins,-kuzzle:services:elasticsearch nodemon bin/start-kuzzle-server --enable-plugins functional-test-plugin,kuzzle-plugin-cluster

# term 3
$ kuzzle_plugins__cluster__privileged=true kuzzle_server__port=8512 kuzzle_plugins__cluster__bindings__router="tcp://[_site_:ipv4]:8510" kuzzle_plugins__cluster__bindings__pub="tcp://[_site_:ipv4]:8511" NODE_ENV=development DEBUG=kuzzle:*,-kuzzle:events*,-kuzzle:entry-point:protocols:websocket,-kuzzle:funnel,-kuzzle:entry-point:protocols:http,-kuzzle:http:router,-kuzzle:plugins,-kuzzle:services:elasticsearch nodemon bin/start-kuzzle-server --enable-plugins functional-test-plugin,kuzzle-plugin-cluster
```

You can then send request to a specific node with Kourou:
```bash
# Send to the first node on port 7512
$ kourou sdk:query realtime:publish -i index -c collection --body '{ hello: "adrien" }' --port 7512

# Send to the second node on port 8512
$ kourou sdk:query realtime:publish -i index -c collection --body '{ hello: "adrien" }' --port 8512
```

### Other changes

 - rename `entryPoints` to `entryPoint`
 - protocol names should be passed as parent constructor argument
 - deprecate `kuzzle:start` and add `kuzzle:start:before` and `kuzzle:start:after`
 - add new `Channel` class

### Boyscout

 - one const per line
 - asyncify some methods
 - use Map instead of object for RAM caches
 - use `Bluebird.all` instead of `Promise.all`