Log ES request errors (#1710)
ES request errors are currently not logged. Most of the time this is not a problem, but having those request logs can be very handy when investigating Kuzzle issues related to document storage.

The log level chosen for this is info so these logs can be easily filtered if they're found to be too verbose on some production environments.