Fix typo in scroll document page (#1761)
## What does this PR do ?

Really short fix of a typo in the documentation, after a try to update the TTL of scroll when you make a search of documents in collection.
I decide to update the documentation for helping people like me who wasn't docker expert and not sure how configurations work. 
I don't arrive to update one value (of course it's little bit more complicated when you don't have the right key 🙃)
So maybe this RP can create a reflexion for improving the kuzzle configuration page or explain how you have the possibility to create settings file for docker image or how to update env variable for docker-compose.yml file.