Make getRequestSlot accept a function in place of a string (#1765)
Make the `getRequestSlot` accept a reference to a function in place of a string representing a function name.
The signature of the function changed from this

```javascript
getRequestSlot (executor, request, executeCallback)
```

to this

```javascript
throttle (fn, context, requestId)
```

The previous implementation assumed that `getRequestSlot` could be called only by methods of the Funnel. The `executor` string represented the name of the method. In fact, `getRequestSlot` was kind of aware that `funnel.execute` was the only function that actually called it.

The new implementation takes a reference to a function and its execution context, as well as a unique id that the funnel uses to keep track of it (the name `requestId` is used because, most of the time, these functions are executed to process incoming requests).