Fix NPM package files (#1767)
Our 2 default plugins kuzzle-plugin-logger and kuzzle-plugin-auth-passport-local are now loaded through NPM and Kaaf.

This was mandatory to distribute the kuzzle package because those 2 plugins were submodules with their own node_modules and it was difficult to include them in the package.

The cluster plugin is not a submodule anymore, all the code is now present in Kuzzle repository. (Existent PR changes must be added when merged on the cluster repo)
All cluster dependencies are Kuzzle dependencies.
This is temporary until we integrate the cluster in the core

Also JSON files were missing from the generated package.