Avoid printing ES errors in development (#1769)
## What does this PR do ?

When ES returns an error, with print it for production debugging purpose.  
The problem is that some errors are "normal" and handled by plugins.  

For example the auth plugin use this mechanism to [create the reset secret](https://github.com/kuzzleio/kuzzle-plugin-auth-passport-local/blob/master/lib/index.js#L212) 

This result in a scary message printed by Kuzzle at startup:
![image](https://user-images.githubusercontent.com/4447392/91841983-8077d880-ec53-11ea-9dfb-ac16026e716e.png)

This PR ensure that ES errors are only logged in production environments.  
The error is also printed as a single JSON string otherwise some information may not be available because they are too deep