Allows to load rights for non-existent API actions at startup (#1775)
## What does this PR do ?

A regression prevent to set rights for API action if those actions does not exists.  

It's problematic because application and plugins can declare new API actions and the only way to prevent unauthorized usage of those action after startup is to set the rights with the `--securities <securities file>` flag with the command line.

Now only a warning will be printed.