Repair functional tests (SDK) (#1815)
# Description

Functional tests (SDK version) weren't launched anymore since a change made on `package.json`, which started the legacy tests instead of the sdk ones.

So, this PR repairs that. And it also repairs the failing functional tests because of the controller naming changes introduced with the new way of adding controller classes.