Allow profile restrictions on Elasticsearch aliases (#1818)
Description
Fix #1807

ES aliases have different naming rules than indexes. When creating, replacing or updating profiles with storage restrictions, Kuzzle verifies if index or collection names are valid. Which prevents some ES aliases to be used in profiles.

This PR changes the following:

by default, no check is done at all on index or collection names, it is the administrators responsbility to ensure that profile rights are correct (worse that can happen: a profile can be restricted to unreachable indexes/collections)
a new strict option is now available to the following routes: security:createProfile, security:createOrReplaceProfile, security:updateProfile. This option is off by default, but if set, it makes Kuzzle verify if provided indexes or collections already exist, otherwise the profile modification is rejected (this is different from the previous behavior, where we only checked that the indexes/collectins could exist)
Other changes
Fix handling of functional tests with "( not)? " sentences: most of them didn't fail if the action succeeded when it should not
As a consequence, fix a failing functional test because of a misconfiguration